app/auth/oauth.py
from fastapi import status, Depends, HTTPException
from jose import JWTError, jwt
from sqlalchemy import select
from sqlalchemy.ext.asyncio import AsyncEngine
from datetime import datetime, timedelta
from fastapi.security import OAuth2PasswordBearer
from database import get_db
